# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Scroungers Multimedia API
  description: |
    Backend API for Scroungers Multimedia platform.
    
    ## Authentication
    Most endpoints require authentication via Supabase Auth. Include the session cookie or Bearer token in requests.
    
    ## Rate Limiting
    Rate limiting is currently disabled.
    
    ## Error Responses
    All errors follow a standard format:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human readable message",
        "details": {}
      }
    }
    ```
  version: 1.0.0
  contact:
    name: Scroungers Multimedia
    url: https://scroungers.co
servers:
  - url: /api
    description: API routes
...
paths:
...
  /auth/callback:
    get:
      tags: [Auth]
      summary: OAuth callback handler
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: next
          in: query
          schema:
            type: string
            default: /dashboard
      responses:
        '307':
          description: Redirect to dashboard or error page

  # =========================================================================
  # CONTACT
  # =========================================================================
...
  /media/presigned-url:
    post:
      tags: [Media]
      summary: Get presigned URL for upload
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, contentType]
              properties:
                filename:
                  type: string
                contentType:
                  type: string
      responses:
        '200':
          description: Presigned URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                  key:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...